modified methods

+++ PlayerMatch.java
@@ -1,11 +1,11 @@
 public class PlayerMatch{
-	public int kills = 0; 
-    public int deaths = 0; 
-    public int assists = 0; 
-    public int kda = 0;
-    public int kd = 0;
-    public int lasthit = 0; 
-    public int denies = 0; 
-    public int gpm = 0; 
-    public int xpm = 0; 
+	public double kills = 0; 
+    public double deaths = 0; 
+    public double assists = 0; 
+    public double kda = 0;
+    public double kd = 0;
+    public double lasthit = 0; 
+    public double denies = 0; 
+    public double gpm = 0; 
+    public double xpm = 0; 
 }
